updated some hardcoded stuff

See merge request CompSci308_2018Spring/voogasalad_iftruereturntrue!375

+++ src/engine/sprites/towers/FrontEndTower.java
@@ -20,7 +20,7 @@ public interface FrontEndTower extends FrontEndSprite{
      * the value of the statistic. NOTE: should include sell price of tower
      * @return
      */
-    public Map<String,Double> getTowerStats();
+    public Map<String,Integer> getTowerStats();
 
     /**
      *
